requirements.txt
Flask==3.1.0
numpy==2.2.4
requests==2.32.3
waitress==3.0.2
six==1.17.0
pandas==2.2.2
geoip2==4.8.0
duckdb
python-dotenv==1.0.1